Refactor comment handling into shared printer module
